backend/requirements.txt
Flask==2.2.5
Flask-RESTX==1.1.0
Flask-Cors==4.0.1
SQLAlchemy==1.4.58
python-dotenv==1.0.0
passlib==1.7.4
PyJWT==2.8.0
email-validator==1.3.1
pillow==10.0.0
